rent/requirements.txt
asgiref==3.4.1
Django==3.2.8
django-crispy-forms==1.13.0
django-modeltranslation==0.17.3
django-phonenumber-field==5.2.0
phonenumbers==8.12.35
phonenumberslite==8.12.35
Pillow==8.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
